+++ ALUMNI/ALUMNI/Common/K.swift
@@ -1,0 +1,22 @@
+//
+//  File.swift
+//  ALUMNI
+//
+//  Created by bushra nazal alatwi on 10/06/1443 AH.
+//
+
+import Foundation
+
+struct k {
+  
+  struct Storyboard {
+    
+    static let homeViewController = "HomeVC"
+    static let newPostVC = "newPost"
+    static let segueGoToPostDetails = "postDetails"
+    static let signInUpVC = "signInUpVC"
+    static let segueGoToChat = "goToChat"
+    
+  }
+  
+}
