+++ ALUMNI/ALUMNI/Helpers/UIView+Ext.swift
@@ -8,6 +8,7 @@
 import UIKit
 
 extension UIView {
+  
     func shakeView() {
         let animation = CABasicAnimation(keyPath: "position")
         animation.duration = 0.09
