+++ ALUMNI/ALUMNI/ViewController/TabBar/MoreVC/AdsVC.swift
@@ -1,5 +1,5 @@
 //
-//  CoursesVC.swift
+//  AdsVC.swift
 //  ALUMNI
 //
 //  Created by bushra nazal alatwi on 29/05/1443 AH.
@@ -8,27 +8,27 @@
 import UIKit
 import Firebase
 
-class CoursesVC: UITableViewController {
+class AdsVC: UITableViewController {
   
-  var courses = [Course]()
-  let db = Firestore.firestore().collection("Courses")
+  var ads = [Ad]()
+  let db = Firestore.firestore().collection("Ads")
   
   
+  //MARK: - View Controller Life Cycle
   override func viewDidLoad() {
     super.viewDidLoad()
     
-    getAllCourses()
+    getAllAds()
   }
-  
   
   @IBAction func newAdAction(_ sender: UIBarButtonItem) {
     
     //1. Create the alert controller.
-    let alert = UIAlertController(title: "Course".localize(), message: "Enter a new course".localize(), preferredStyle: .alert)
+    let alert = UIAlertController(title: "Ads".localize(), message: "Enter a new Ad".localize(), preferredStyle: .alert)
     
     //2. Add the text field. You can configure it however you need.
     alert.addTextField { (textField) in
-      textField.placeholder = "Course Title".localize()
+      textField.placeholder = "Ad Title".localize()
     }
     
     // 3. Grab the value from the text field, and print it when the user clicks OK.
@@ -36,9 +36,9 @@ class CoursesVC: UITableViewController {
       let textField = alert?.textFields![0] // Force unwrapping because we know it exists.
       guard let text = textField?.text, text.isEmpty == false else {return}
       guard let userID = Auth.auth().currentUser?.uid else {return}
-      let courseID = UUID().uuidString
-      self.db.document(courseID).setData([
-        "courseID" : courseID,
+      let adID = UUID().uuidString
+      self.db.document(adID).setData([
+        "adID" : adID,
         "text" : text,
         "userID" : userID,
         "timestamp" : Date().timeIntervalSince1970
@@ -52,16 +52,16 @@ class CoursesVC: UITableViewController {
     
   }
   
-  func getAllCourses() {
+  func getAllAds() {
     db.order(by: "timestamp", descending: true).addSnapshotListener { snapshot, error in
-      self.courses.removeAll()
+      self.ads.removeAll()
       if let value = snapshot?.documents {
         for i in value {
           let data = i.data()
-          let adID = data["courseID"] as? String
+          let adID = data["adID"] as? String
           let text = data["text"] as? String
           let userID = data["userID"] as? String
-          self.courses.append(Course(courseID: adID, text: text, userID: userID))
+          self.ads.append(Ad(adID: adID, text: text, userID: userID))
         }
         self.tableView.reloadData()
       }
@@ -69,23 +69,20 @@ class CoursesVC: UITableViewController {
   }
   
   
-  
-  
   // MARK: - Table view data source
+  
   override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
     // #warning Incomplete implementation, return the number of rows
-    return courses.count
+    return ads.count
   }
   
   
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
     let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
     
-    cell.textLabel?.text = courses[indexPath.row].text
+    cell.textLabel?.text = ads[indexPath.row].text
     
     return cell
   }
   
-  
 }
-
