+++ ALUMNI/ALUMNI/ViewController/TabBar/MoreVC/CoursesVC.swift
@@ -1,5 +1,5 @@
 //
-//  AdsVC.swift
+//  CoursesVC.swift
 //  ALUMNI
 //
 //  Created by bushra nazal alatwi on 29/05/1443 AH.
@@ -8,25 +8,28 @@
 import UIKit
 import Firebase
 
-class AdsVC: UITableViewController {
+class CoursesVC: UITableViewController {
   
-  var ads = [Ad]()
-  let db = Firestore.firestore().collection("Ads")
+  var courses = [Course]()
+  let db = Firestore.firestore().collection("Courses")
   
+  
+  //MARK: - View Controller Life Cycle
   override func viewDidLoad() {
     super.viewDidLoad()
     
-    getAllAds()
+    getAllCourses()
   }
+  
   
   @IBAction func newAdAction(_ sender: UIBarButtonItem) {
     
     //1. Create the alert controller.
-    let alert = UIAlertController(title: "Ads".localize(), message: "Enter a new Ad".localize(), preferredStyle: .alert)
+    let alert = UIAlertController(title: "Course".localize(), message: "Enter a new course".localize(), preferredStyle: .alert)
     
     //2. Add the text field. You can configure it however you need.
     alert.addTextField { (textField) in
-      textField.placeholder = "Ad Title".localize()
+      textField.placeholder = "Course Title".localize()
     }
     
     // 3. Grab the value from the text field, and print it when the user clicks OK.
@@ -34,9 +37,9 @@ class AdsVC: UITableViewController {
       let textField = alert?.textFields![0] // Force unwrapping because we know it exists.
       guard let text = textField?.text, text.isEmpty == false else {return}
       guard let userID = Auth.auth().currentUser?.uid else {return}
-      let adID = UUID().uuidString
-      self.db.document(adID).setData([
-        "adID" : adID,
+      let courseID = UUID().uuidString
+      self.db.document(courseID).setData([
+        "courseID" : courseID,
         "text" : text,
         "userID" : userID,
         "timestamp" : Date().timeIntervalSince1970
@@ -50,16 +53,16 @@ class AdsVC: UITableViewController {
     
   }
   
-  func getAllAds() {
+  func getAllCourses() {
     db.order(by: "timestamp", descending: true).addSnapshotListener { snapshot, error in
-      self.ads.removeAll()
+      self.courses.removeAll()
       if let value = snapshot?.documents {
         for i in value {
           let data = i.data()
-          let adID = data["adID"] as? String
+          let adID = data["courseID"] as? String
           let text = data["text"] as? String
           let userID = data["userID"] as? String
-          self.ads.append(Ad(adID: adID, text: text, userID: userID))
+          self.courses.append(Course(courseID: adID, text: text, userID: userID))
         }
         self.tableView.reloadData()
       }
@@ -67,20 +70,23 @@ class AdsVC: UITableViewController {
   }
   
   
-  // MARK: - Table view data source
   
+  
+  // MARK: - Table view data source
   override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
     // #warning Incomplete implementation, return the number of rows
-    return ads.count
+    return courses.count
   }
   
   
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
     let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
     
-    cell.textLabel?.text = ads[indexPath.row].text
+    cell.textLabel?.text = courses[indexPath.row].text
     
     return cell
   }
   
+  
 }
+
